requirements.txt
opencv-python
numpy
pillow
streamlit
google-generativeai
python-dotenv